+++ test/matchers/have_field_test.rb
@@ -5,4 +5,7 @@ describe "HaveFieldMatcher" do
 
   it { must have_field(:name) }
   it { wont have_field(:noexist) }
+
+  it { must have_field(:name).of_type(String) }
+  it { wont have_field(:name).of_type(Integer) }
 end