+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.1.1
 
+* Adding `of_type` validation to HaveFieldMatcher.
 * Adding `must` and `wont` to MiniTest::Spec to use them with subject.
 * Adding `have_field` matcher.
 * Adding `must` and `wont` expectations.